atividades/requirements.txt
Flask-SQLAlchemy
Flask-Swagger-UI
requests
